requirements.txt
Django==5.0
djangorestframework==3.14.0
django-filter==23.5
Pillow==10.1.0
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
numpy==1.26.2
scikit-learn==1.3.2
pandas==2.1.4
djangorestframework-simplejwt==5.3.1